Refactored ghost code
