+++ lua/autorun/client/advdupe2_cl_init.lua
@@ -15,6 +15,7 @@ include "advdupe2/file_browser.lua"
 include "advdupe2/sh_codec.lua"
 include "advdupe2/sh_netstream.lua"
 include "advdupe2/cl_file.lua"
+include "advdupe2/cl_ghost.lua"
 
 function AdvDupe2.Notify(msg,typ,dur)
 	surface.PlaySound(typ == 1 and "buttons/button10.wav" or "ambient/water/drip1.wav")
