+++ lua/weapons/gmod_tool/stools/advdupe2.lua
@@ -1,10 +1,10 @@
 --[[
 	Title: Adv. Dupe 2 Tool
-	
+
 	Desc: Defines the AD2 tool and assorted functionalities.
-	
+
 	Author: TB
-	
+
 	Version: 1.0
 ]]
 TOOL.Category = "Construction"
@@ -53,7 +53,7 @@ if(SERVER)then
 					for systemi, system in pairs(sortingSystems) do
 						for _, target in pairs(system) do
 							for x = 1, 4 do
-								if v.Entity[x] then 
+								if v.Entity[x] then
 									for y = 1, 4 do
 										if target.Entity[y] and v.Entity[x].Index == target.Entity[y].Index then
 											system[#system + 1] = v
@@ -115,7 +115,7 @@ if(SERVER)then
 		end
 		return ret
 	end
-	
+
 	local function GetSortedConstraints( ply, constraints )
 		if ply:GetInfo("advdupe2_sort_constraints") ~= 0 then
 			return GroupConstraintOrder( ply, constraints )
@@ -138,7 +138,7 @@ if(SERVER)then
 		local trace = WireLib and WireLib.dummytrace(ent) or { Entity = ent }
 		return hook.Run( "CanTool", ply,  trace, "advdupe2" ) ~= false
 	end
-	
+
 	--Find all the entities in a box, given the adjacent corners and the player
 	local function FindInBox(min, max, ply)
 		local PPCheck = (tobool(ply:GetInfo("advdupe2_copy_only_mine")) and CPPI~=nil) and PlayerCanDupeCPPI or PlayerCanDupeTool
@@ -148,14 +148,14 @@ if(SERVER)then
 		for i=1, #Entities do
 			ent = Entities[i]
 			pos = ent:GetPos()
-			if (pos.X>=min.X) and (pos.X<=max.X) and (pos.Y>=min.Y) and (pos.Y<=max.Y) and (pos.Z>=min.Z) and (pos.Z<=max.Z) and PPCheck( ply, ent ) then	
+			if (pos.X>=min.X) and (pos.X<=max.X) and (pos.Y>=min.Y) and (pos.Y<=max.Y) and (pos.Z>=min.Z) and (pos.Z<=max.Z) and PPCheck( ply, ent ) then
 				EntTable[ent:EntIndex()] = ent
 			end
 		end
 
 		return EntTable
 	end
-	
+
 	--[[
 		Name: LeftClick
 		Desc: Defines the tool's behavior when the player left-clicks.
@@ -167,17 +167,17 @@ if(SERVER)then
 
 		local ply = self:GetOwner()
 		if(not ply.AdvDupe2 or not ply.AdvDupe2.Entities)then return false end
-		
+
 		if(ply.AdvDupe2.Pasting or ply.AdvDupe2.Downloading)then
 			AdvDupe2.Notify(ply,"Advanced Duplicator 2 is busy.",NOTIFY_ERROR)
-			return false 
+			return false
 		end
-		
+
 		local z = math.Clamp((tonumber(ply:GetInfo("advdupe2_offset_z")) + ply.AdvDupe2.HeadEnt.Z), -16000, 16000)
 		ply.AdvDupe2.Position = trace.HitPos + Vector(0, 0, z)
 		ply.AdvDupe2.Angle = Angle(ply:GetInfoNum("advdupe2_offset_pitch", 0), ply:GetInfoNum("advdupe2_offset_yaw", 0), ply:GetInfoNum("advdupe2_offset_roll", 0))
 		if(tobool(ply:GetInfo("advdupe2_offset_world")))then ply.AdvDupe2.Angle = ply.AdvDupe2.Angle - ply.AdvDupe2.Entities[ply.AdvDupe2.HeadEnt.Index].PhysicsObjects[0].Angle end
-		
+
 		ply.AdvDupe2.Pasting = true
 		AdvDupe2.Notify(ply,"Pasting...")
 		local origin
@@ -187,7 +187,7 @@ if(SERVER)then
 		AdvDupe2.InitPastingQueue(ply, ply.AdvDupe2.Position, ply.AdvDupe2.Angle, origin, tobool(ply:GetInfo("advdupe2_paste_constraints")), tobool(ply:GetInfo("advdupe2_paste_parents")), tobool(ply:GetInfo("advdupe2_paste_disparents")),tobool(ply:GetInfo("advdupe2_paste_protectoveride")))
 		return true
 	end
-	
+
 	--[[
 		Name: RightClick
 		Desc: Defines the tool's behavior when the player right-clicks.
@@ -196,10 +196,10 @@ if(SERVER)then
 	]]
 	function TOOL:RightClick( trace )
 		local ply = self:GetOwner()
-		
+
 		if(ply.AdvDupe2.Pasting or ply.AdvDupe2.Downloading)then
 			AdvDupe2.Notify(ply,"Advanced Duplicator 2 is busy.", NOTIFY_ERROR)
-			return false 
+			return false
 		end
 
 		--Set Area Copy on or off
@@ -212,11 +212,11 @@ if(SERVER)then
 				AdvDupe2.RemoveSelectBox(ply)
 				self:SetStage(0)
 				return false
-			end	
+			end
 		end
-		
+
 		if(not trace or not trace.Hit)then return false end
-		
+
 		local Entities, Constraints, AddOne
 		local HeadEnt = {}
 		--If area copy is on
@@ -224,23 +224,23 @@ if(SERVER)then
 			local area_size = tonumber(ply:GetInfo("advdupe2_area_copy_size")) or 50
 			if( not game.SinglePlayer() and area_size > tonumber(GetConVarString("AdvDupe2_MaxAreaCopySize")))then
 				AdvDupe2.Notify(ply,"Area copy size exceeds limit of "..GetConVarString("AdvDupe2_MaxAreaCopySize")..".",NOTIFY_ERROR)
-				return false 
+				return false
 			end
 			local Pos = trace.HitNonWorld and trace.Entity:GetPos() or trace.HitPos
 			local T = (Vector(area_size,area_size,area_size)+Pos)
 			local B = (Vector(-area_size,-area_size,-area_size)+Pos)
-			
+
 			local Ents = FindInBox(B,T, ply)
 			if next(Ents)==nil then
 				self:SetStage(0)
 				AdvDupe2.RemoveSelectBox(ply)
 				return true
 			end
-			
+
 			local Ent = trace.HitNonWorld and trace.Entity or Ents[next(Ents)]
 			HeadEnt.Index = Ent:EntIndex()
 			HeadEnt.Pos = Ent:GetPos()
-			
+
 			Entities, Constraints = AdvDupe2.duplicator.AreaCopy(Ents, HeadEnt.Pos, tobool(ply:GetInfo("advdupe2_copy_outside")))
 
 			self:SetStage(0)
@@ -256,9 +256,9 @@ if(SERVER)then
 				Entities = ply.AdvDupe2.Entities
 				Constraints = ply.AdvDupe2.Constraints
 				HeadEnt = ply.AdvDupe2.HeadEnt
-				
+
 				AdvDupe2.duplicator.Copy( trace.Entity, Entities, Constraints, HeadEnt.Pos)
-				
+
 				--Only add the one ghost
 				AddOne = Entities[trace.Entity:EntIndex()]
 			else
@@ -266,7 +266,7 @@ if(SERVER)then
 				Constraints = {}
 				HeadEnt.Index = trace.Entity:EntIndex()
 				HeadEnt.Pos = trace.HitPos
-				
+
 				AdvDupe2.duplicator.Copy( trace.Entity, Entities, Constraints, trace.HitPos )
 			end
 		else --Non valid entity or clicked the world
@@ -294,24 +294,24 @@ if(SERVER)then
 					umsg.End()
 					return true
 				end
-				
+
 				local Ent = Entities[next(Entities)]
 				HeadEnt.Index = Ent:EntIndex()
 				HeadEnt.Pos = Ent:GetPos()
-				
+
 				Entities, Constraints = AdvDupe2.duplicator.AreaCopy(Entities, HeadEnt.Pos, tobool(ply:GetInfo("advdupe2_copy_outside")))
 			end
 		end
-		
+
 		if not HeadEnt.Z then
 			local WorldTrace = util.TraceLine( {mask=MASK_NPCWORLDSTATIC, start=HeadEnt.Pos+Vector(0,0,1), endpos=HeadEnt.Pos-Vector(0,0,50000)} )
 			HeadEnt.Z = WorldTrace.Hit and math.abs(HeadEnt.Pos.Z-WorldTrace.HitPos.Z) or 0
 		end
-		
+
 		ply.AdvDupe2.HeadEnt = HeadEnt
 		ply.AdvDupe2.Entities = Entities
 		ply.AdvDupe2.Constraints = GetSortedConstraints(ply, Constraints)
-		
+
 		net.Start("AdvDupe2_SetDupeInfo")
 			net.WriteString("")
 			net.WriteString(ply:Nick())
@@ -324,38 +324,30 @@ if(SERVER)then
 		net.Send(ply)
 
 		if AddOne then
-			net.Start("AdvDupe2_AddGhost")
-				net.WriteBit(AddOne.Class=="prop_ragdoll")
-				net.WriteString(AddOne.Model)
-				net.WriteInt(#AddOne.PhysicsObjects, 8)
-				for i=0, #AddOne.PhysicsObjects do
-					net.WriteAngle(AddOne.PhysicsObjects[i].Angle)
-					net.WriteVector(AddOne.PhysicsObjects[i].Pos)
-				end
-			net.Send(ply)
+			AdvDupe2.SendGhost(ply, AddOne)
 		else
-			AdvDupe2.SendGhosts(ply) 
+			AdvDupe2.SendGhosts(ply)
 		end
 
 		AdvDupe2.ResetOffsets(ply)
 
 		return true
 	end
-	
+
 	--Checks table, re-draws loading bar, and recreates ghosts when tool is pulled out
 	function TOOL:Deploy()
 		local ply = self:GetOwner()
-		
+
 		if(not ply.AdvDupe2.Entities)then return end
-		
+
 		umsg.Start("AdvDupe2_StartGhosting", ply)
 		umsg.End()
-		
+
 		if(ply.AdvDupe2.Queued)then
 			AdvDupe2.InitProgressBar(ply, "Queued: ")
 			return
 		end
-		
+
 		if(ply.AdvDupe2.Pasting)then
 			AdvDupe2.InitProgressBar(ply, "Pasting: ")
 			return
@@ -384,13 +376,13 @@ if(SERVER)then
 	]]
 	function TOOL:Reload( trace )
 		if(!trace.Hit)then return false end
-		
+
 		local ply = self:GetOwner()
-		
+
 		if(self:GetStage()==1)then
 			if( not game.SinglePlayer() and (tonumber(ply:GetInfo("advdupe2_area_copy_size"))or 50) > tonumber(GetConVarString("AdvDupe2_MaxAreaCopySize")))then
 				AdvDupe2.Notify(ply,"Area copy size exceeds limit of "..GetConVarString("AdvDupe2_MaxAreaCopySize")..".",NOTIFY_ERROR)
-				return false 
+				return false
 			end
 			umsg.Start("AdvDupe2_CanAutoSave", ply)
 				umsg.Vector(trace.HitPos)
@@ -408,7 +400,7 @@ if(SERVER)then
 			ply.AdvDupe2.TempAutoSaveOutSide = tobool(ply:GetInfo("advdupe2_copy_outside"))
 			return true
 		end
-		
+
 		--If a contraption spawner was clicked then update it with the current settings
 		if(trace.Entity:GetClass()=="gmod_contr_spawner")then
 			local delay = tonumber(ply:GetInfo("advdupe2_contr_spawner_delay"))
@@ -427,7 +419,7 @@ if(SERVER)then
 					delay = 0
 				end
 			end
-			
+
 			if(not undo_delay)then
 				undo_delay = tonumber(GetConVarString("AdvDupe2_MinContraptionUndoDelay"))
 			else
@@ -451,7 +443,7 @@ if(SERVER)then
 		if ply.AdvDupe2 and ply.AdvDupe2.Entities then
 			local headent = ply.AdvDupe2.Entities[ply.AdvDupe2.HeadEnt.Index]
 			local Pos, Ang
-			
+
 			if(headent)then
 				if(tobool(ply:GetInfo("advdupe2_original_origin")))then
 					Pos = ply.AdvDupe2.HeadEnt.Pos + headent.PhysicsObjects[0].Pos
@@ -460,19 +452,19 @@ if(SERVER)then
 					local EntAngle = headent.PhysicsObjects[0].Angle
 					if(tobool(ply:GetInfo("advdupe2_offset_world")))then EntAngle = Angle(0,0,0) end
 					trace.HitPos.Z = trace.HitPos.Z + math.Clamp(ply.AdvDupe2.HeadEnt.Z + tonumber(ply:GetInfo("advdupe2_offset_z")) or 0, -16000, 16000)
-					Pos, Ang = LocalToWorld(headent.PhysicsObjects[0].Pos, EntAngle, trace.HitPos, Angle(math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_pitch")) or 0,-180,180), math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_yaw")) or 0,-180,180), math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_roll")) or 0,-180,180))) 
+					Pos, Ang = LocalToWorld(headent.PhysicsObjects[0].Pos, EntAngle, trace.HitPos, Angle(math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_pitch")) or 0,-180,180), math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_yaw")) or 0,-180,180), math.Clamp(tonumber(ply:GetInfo("advdupe2_offset_roll")) or 0,-180,180)))
 				end
 			else
 				AdvDupe2.Notify(ply, "Invalid head entity to spawn contraption spawner.")
 				return false
 			end
-			
-			if(headent.Class=="gmod_contr_spawner") then 
+
+			if(headent.Class=="gmod_contr_spawner") then
 				AdvDupe2.Notify(ply, "Cannot make a contraption spawner from a contraption spawner.")
-				return false 
+				return false
 			end
-		
-			
+
+
 			local spawner = MakeContraptionSpawner( ply, Pos, Ang, ply.AdvDupe2.HeadEnt.Index, table.Copy(ply.AdvDupe2.Entities), table.Copy(ply.AdvDupe2.Constraints), tonumber(ply:GetInfo("advdupe2_contr_spawner_delay")), tonumber(ply:GetInfo("advdupe2_contr_spawner_undo_delay")), headent.Model, tonumber(ply:GetInfo("advdupe2_contr_spawner_key")), tonumber(ply:GetInfo("advdupe2_contr_spawner_undo_key")),  tonumber(ply:GetInfo("advdupe2_contr_spawner_disgrav")) or 0, tonumber(ply:GetInfo("advdupe2_contr_spawner_disdrag")) or 0, tonumber(ply:GetInfo("advdupe2_contr_spawner_addvel")) or 1, tonumber(ply:GetInfo("advdupe2_contr_spawner_hideprops")) or 0 )
 			ply:AddCleanup( "AdvDupe2", spawner )
 			undo.Create("gmod_contr_spawner")
@@ -482,7 +474,7 @@ if(SERVER)then
 
 			return true
 		end
-	end	
+	end
 
 	--Called to clean up the tool when pasting is finished or undo during pasting
 	function AdvDupe2.FinishPasting(Player, Paste)
@@ -490,77 +482,20 @@ if(SERVER)then
 		AdvDupe2.RemoveProgressBar(Player)
 		if(Paste)then AdvDupe2.Notify(Player,"Finished Pasting!") end
 	end
-	
-	function AdvDupe2.SendGhosts(ply)
-		if(not ply.AdvDupe2.Entities)then return end
-		
-		local cache = {}
-		local temp = {}
-		local mdls = {}
-		local cnt = 1
-		local add = true
-		local head
-
-		for k,v in pairs(ply.AdvDupe2.Entities)do
-			temp[cnt] = v
-			for i=1,#cache do
-				if(cache[i]==v.Model)then
-					mdls[cnt] = i
-					add=false
-					break
-				end
-			end
-			if(add)then
-				mdls[cnt] = table.insert(cache, v.Model)
-			else
-				add = true
-			end
-			if(k==ply.AdvDupe2.HeadEnt.Index)then
-				head = cnt
-			end
-			cnt = cnt+1
-		end
-		
-		if(!head)then
-			AdvDupe2.Notify(ply, "Invalid head entity for ghosts.", NOTIFY_ERROR);
-			return
-		end
-		
-		net.Start("AdvDupe2_SendGhosts")
-			net.WriteInt(head, 16)
-			net.WriteFloat(ply.AdvDupe2.HeadEnt.Z)
-			net.WriteVector(ply.AdvDupe2.HeadEnt.Pos)
-			net.WriteInt(#cache, 16)
-			for i=1,#cache do
-				net.WriteString(cache[i])
-			end
-			net.WriteInt(cnt-1, 16)
-			for i=1, #temp do
-				net.WriteBit(temp[i].Class=="prop_ragdoll")
-				net.WriteInt(mdls[i], 16)
-				net.WriteInt(#temp[i].PhysicsObjects, 8)
-				for k=0, #temp[i].PhysicsObjects do
-					net.WriteAngle(temp[i].PhysicsObjects[k].Angle)
-					net.WriteVector(temp[i].PhysicsObjects[k].Pos)
-				end
-			end
-		net.Send(ply)
-		
-	end
 
 	--function for creating a contraption spawner
 	function MakeContraptionSpawner( ply, Pos, Ang, HeadEnt, EntityTable, ConstraintTable, delay, undo_delay, model, key, undo_key, disgrav, disdrag, addvel, hideprops)
 
 		if not ply:CheckLimit("gmod_contr_spawners") then return nil end
-		
+
 		if(not game.SinglePlayer())then
 			if(table.Count(EntityTable)>tonumber(GetConVarString("AdvDupe2_MaxContraptionEntities")))then
 				AdvDupe2.Notify(ply,"Contraption Spawner exceeds the maximum amount of "..GetConVarString("AdvDupe2_MaxContraptionEntities").." entities for a spawner!",NOTIFY_ERROR)
-				return false 
+				return false
 			end
 			if(#ConstraintTable>tonumber(GetConVarString("AdvDupe2_MaxContraptionConstraints")))then
 				AdvDupe2.Notify(ply,"Contraption Spawner exceeds the maximum amount of "..GetConVarString("AdvDupe2_MaxContraptionConstraints").." constraints for a spawner!",NOTIFY_ERROR)
-				return false 
+				return false
 			end
 		end
 
@@ -574,7 +509,7 @@ if(SERVER)then
 		spawner:Spawn()
 
 		duplicator.ApplyEntityModifiers(ply, spawner)
-		
+
 		if IsValid(spawner:GetPhysicsObject()) then
 			spawner:GetPhysicsObject():EnableMotion(false)
 		end
@@ -593,7 +528,7 @@ if(SERVER)then
 				delay = 0
 			end
 		end
-		
+
 		if(not undo_delay)then
 			undo_delay = tonumber(GetConVarString("AdvDupe2_MinContraptionUndoDelay"))
 		else
@@ -609,7 +544,7 @@ if(SERVER)then
 				undo_delay = 0
 			end
 		end
-			
+
 		-- Set options
 		spawner:SetPlayer(ply)
 		spawner:GetTable():SetOptions(ply, delay, undo_delay, key, undo_key, disgrav, disdrag, addvel, hideprops)
@@ -633,7 +568,7 @@ if(SERVER)then
 	end
 	duplicator.RegisterEntityClass("gmod_contr_spawner", MakeContraptionSpawner, "Pos", "Ang", "HeadEnt", "EntityTable", "ConstraintTable", "delay", "undo_delay", "model", "key", "undo_key", "disgrav", "disdrag", "addvel", "hideprops")
 
-	
+
 	function AdvDupe2.InitProgressBar(ply,label)
 		umsg.Start("AdvDupe2_InitProgressBar",ply)
 			umsg.String(label)
@@ -663,7 +598,7 @@ if(SERVER)then
 
 	--Reset the offsets of height, pitch, yaw, and roll back to default
 	function AdvDupe2.ResetOffsets(ply, keep)
-		
+
 		if(not keep)then
 			ply.AdvDupe2.Name = nil
 		end
@@ -676,9 +611,9 @@ if(SERVER)then
 			umsg.Short(perc)
 		umsg.End()
 	end
-	
+
 	net.Receive("AdvDupe2_CanAutoSave", function(len, ply, len2)
-	
+
 		local desc = net.ReadString()
 		local ent = net.ReadInt(16)
 
@@ -694,13 +629,13 @@ if(SERVER)then
 			end
 			ply.AdvDupe2.AutoSaveEnt = nil
 		end
-		
+
 		ply.AdvDupe2.AutoSavePos = ply.AdvDupe2.TempAutoSavePos
 		ply.AdvDupe2.AutoSaveSize = ply.AdvDupe2.TempAutoSaveSize
 		ply.AdvDupe2.AutoSaveOutSide = ply.AdvDupe2.TempAutoSaveOutSide
 		ply.AdvDupe2.AutoSaveContr = ply:GetInfo("advdupe2_auto_save_contraption")=="1"
 		ply.AdvDupe2.AutoSaveDesc = desc
-		
+
 		local time = tonumber(ply:GetInfo("advdupe2_auto_save_time")) or 5
 		if(game.SinglePlayer())then
 			ply.AdvDupe2.AutoSavePath = net.ReadString()
@@ -708,52 +643,52 @@ if(SERVER)then
 			if(time>30)then time = 30 end
 			if(time<GetConVarNumber("AdvDupe2_AreaAutoSaveTime"))then time = GetConVarNumber("AdvDupe2_AreaAutoSaveTime") end
 		end
-		
+
 		AdvDupe2.Notify(ply, "Your area will be auto saved every "..(time*60).." seconds.")
 		local name = "AdvDupe2_AutoSave_"..ply:UniqueID()
 		if(timer.Exists(name))then
 			timer.Adjust(name, time*60, 0)
-			return 
+			return
 		end
 		timer.Create(name, time*60, 0, function()
 			if(not IsValid(ply))then
 				timer.Remove(name)
 				return
 			end
-			
+
 			if(ply.AdvDupe2.Downloading)then
 				AdvDupe2.Notify(ply, "Skipping auto save, tool is busy.", NOTIFY_ERROR)
 				return
 			end
-			
+
 			local Tab = {Entities={}, Constraints={}, HeadEnt={}}
-			
+
 			if(ply.AdvDupe2.AutoSaveContr)then
 				if(not IsValid(ply.AdvDupe2.AutoSaveEnt))then
 					timer.Remove(name)
 					AdvDupe2.Notify(ply, "Head entity for auto save no longer valid; stopping auto save.", NOTIFY_ERROR)
 					return
 				end
-				
+
 				Tab.HeadEnt.Index = ply.AdvDupe2.AutoSaveEnt:EntIndex()
 				Tab.HeadEnt.Pos = ply.AdvDupe2.AutoSaveEnt:GetPos()
-				
+
 				local WorldTrace = util.TraceLine( {mask=MASK_NPCWORLDSTATIC, start=Tab.HeadEnt.Pos+Vector(0,0,1), endpos=Tab.HeadEnt.Pos-Vector(0,0,50000)} )
 				if(WorldTrace.Hit)then Tab.HeadEnt.Z = math.abs(Tab.HeadEnt.Pos.Z-WorldTrace.HitPos.Z) else Tab.HeadEnt.Z = 0 end
-				
+
 				AdvDupe2.duplicator.Copy( ply.AdvDupe2.AutoSaveEnt, Tab.Entities, Tab.Constraints, Tab.HeadEnt.Pos )
 			else
 				local i = ply.AdvDupe2.AutoSaveSize
 				local Pos = ply.AdvDupe2.AutoSavePos
 				local T = (Vector(i,i,i)+Pos)
 				local B = (Vector(-i,-i,-i)+Pos)
-				
+
 				local Entities = FindInBox(B,T, ply)
 				if(table.Count(Entities)==0)then
 					AdvDupe2.Notify(ply, "Area Auto Save copied 0 entities; be sure to turn it off.", NOTIFY_ERROR)
 					return
 				end
-				
+
 				if(ply.AdvDupe2.AutoSaveEnt && Entities[ply.AdvDupe2.AutoSaveEnt])then
 					Tab.HeadEnt.Index = ply.AdvDupe2.AutoSaveEnt
 				else
@@ -776,30 +711,30 @@ if(SERVER)then
 		end)
 		timer.Start(name)
 	end)
-	
+
 	concommand.Add("AdvDupe2_SetStage", function(ply, cmd, args)
 		ply:GetTool("advdupe2"):SetStage(1)
 	end)
-	
+
 	concommand.Add("AdvDupe2_RemoveAutoSave", function(ply, cmd, args)
 		timer.Remove("AdvDupe2_AutoSave_"..ply:UniqueID())
 	end)
-	
+
 	concommand.Add("AdvDupe2_SaveMap", function(ply, cmd, args)
 		if(not ply:IsAdmin())then
 			AdvDupe2.Notify(ply, "You do not have permission to this function.", NOTIFY_ERROR)
 			return
 		end
-		
+
 		local Entities = ents.GetAll()
 		for k,v in pairs(Entities) do
 			if v:CreatedByMap() or not duplicator.IsAllowed(v:GetClass()) then
 				Entities[k]=nil
 			end
 		end
-		
+
 		if(table.Count(Entities)==0)then return end
-		
+
 		local Tab = {Entities={}, Constraints={}, HeadEnt={}, Description=""}
 		Tab.HeadEnt.Index = table.GetFirstKey(Entities)
 		Tab.HeadEnt.Pos = Entities[Tab.HeadEnt.Index]:GetPos()
@@ -808,13 +743,13 @@ if(SERVER)then
 		if(WorldTrace.Hit)then Tab.HeadEnt.Z = math.abs(Tab.HeadEnt.Pos.Z-WorldTrace.HitPos.Z) else Tab.HeadEnt.Z = 0 end
 		Tab.Entities, Tab.Constraints = AdvDupe2.duplicator.AreaCopy(Entities, Tab.HeadEnt.Pos, true)
 		Tab.Constraints = GetSortedConstraints(ply, Tab.Constraints)
-		
+
 		Tab.Map = true
 		AdvDupe2.Encode( Tab, AdvDupe2.GenerateDupeStamp(ply), function(data)
 			if(not file.IsDir("advdupe2_maps", "DATA"))then
 				file.CreateDir("advdupe2_maps")
 			end
-			file.Write("advdupe2_maps/"..args[1]..".txt", data)	
+			file.Write("advdupe2_maps/"..args[1]..".txt", data)
 			AdvDupe2.Notify(ply, "Map save, saved successfully.")
 		end)
 	end)
@@ -828,14 +763,14 @@ if(CLIENT)then
 		end
 		return false
 	end
-	
+
 	function TOOL:RightClick(trace)
 		if( self:GetOwner():KeyDown(IN_SPEED) and not self:GetOwner():KeyDown(IN_WALK) )then
 			return false
 		end
 		return true
 	end
-	
+
 	--Removes progress bar and removes ghosts when tool is put away
 	function TOOL:Holster()
 		AdvDupe2.RemoveGhosts()
@@ -843,9 +778,9 @@ if(CLIENT)then
 		if(AdvDupe2.Rotation)then
 			hook.Remove("PlayerBindPress", "AdvDupe2_BindPress")
 		end
-		return 
+		return
 	end
-	
+
 	function TOOL:Reload( trace )
 		if(trace and (AdvDupe2.HeadGhost || self:GetStage()==1))then
 			return true
@@ -877,10 +812,10 @@ if(CLIENT)then
 			end
 			return true
 		end
-		
+
 		GAMEMODE:PlayerBindPress(ply, bind, pressed)
 	end
-	
+
 	local XTotal = 0
 	local YTotal = 0
 	local LastXDegree = 0
@@ -890,14 +825,14 @@ if(CLIENT)then
 
 		local X2 = 0
 		local Y2 = 0
-		
-		if(X~=0)then	
+
+		if(X~=0)then
 			X2 = tonumber(LocalPlayer():GetInfo("advdupe2_offset_yaw"))
-			
+
 			if(LocalPlayer():KeyDown(IN_SPEED))then
 				XTotal = XTotal + X
 				local temp = XTotal + X2
-				
+
 				local degree = math.Round(temp/45)*45
 				if(degree>=225)then
 					degree = -135
@@ -908,7 +843,7 @@ if(CLIENT)then
 					XTotal = 0
 					LastXDegree = degree
 				end
-				
+
 				X2 = degree
 			else
 				X2 = X2 + X
@@ -920,69 +855,43 @@ if(CLIENT)then
 			end
 			RunConsoleCommand("advdupe2_offset_yaw", X2)
 		end
-		
+
 		/*if(Y~=0)then
 			local modyaw = LocalPlayer():GetAngles().y
 			local modyaw2 = tonumber(LocalPlayer():GetInfo("advdupe2_offset_yaw"))
-			
+
 			if(modyaw<0)then modyaw = modyaw + 360 else modyaw = modyaw + 180 end
 			if(modyaw2<0)then modyaw2 = modyaw2 + 360 else modyaw2 = modyaw2 + 180 end
-			
+
 			modyaw = modyaw - modyaw2
 			local modyaw3 = modyaw
 			if(modyaw3<0)then
 				modyaw3 = modyaw3 * -1
 			end
-			
+
 			local pitch = tonumber(LocalPlayer():GetInfo("advdupe2_offset_pitch"))
 			local roll = tonumber(LocalPlayer():GetInfo("advdupe2_offset_roll"))
-			
+
 				--print(modyaw3)
 			if(modyaw3 <= 90)then
 				pitch = pitch + (Y - Y * (modyaw3/90))
 				roll = roll - (Y*(modyaw3/90))
 			end
-			
+
 			--if(pitch>180)then pitch = -180
-			
+
 			RunConsoleCommand("advdupe2_offset_pitch",pitch)
 			RunConsoleCommand("advdupe2_offset_roll",roll)
 		end*/
-		
-	end
-
-	--Update the ghost's postion and angles based on where the player is looking and the offsets
-	local Utrace, UGhostEnt, UEntAngle, UPos, UAngle
-	local function UpdateGhost()
-		Utrace = util.TraceLine(util.GetPlayerTrace(LocalPlayer(), LocalPlayer():GetAimVector()))
-		if (not Utrace.Hit) then return end
-
-		UGhostEnt = AdvDupe2.HeadGhost
-		
-		if(not IsValid(UGhostEnt))then
-			AdvDupe2.RemoveGhosts()
-			AdvDupe2.Notify("Invalid ghost parent.", NOTIFY_ERROR)
-			return 
-		end
-		
-		if(tobool(GetConVarNumber("advdupe2_original_origin")))then
-			UGhostEnt:SetPos(AdvDupe2.HeadPos + AdvDupe2.HeadOffset)
-			UGhostEnt:SetAngles(AdvDupe2.HeadAngle)
-		else
-			UEntAngle = AdvDupe2.HeadAngle
-			if(tobool(GetConVarNumber("advdupe2_offset_world")))then UEntAngle = Angle(0,0,0) end
-			Utrace.HitPos.Z = Utrace.HitPos.Z + math.Clamp(AdvDupe2.HeadZPos + GetConVarNumber("advdupe2_offset_z") or 0, -16000, 16000)
-			UPos, UAngle = LocalToWorld(AdvDupe2.HeadOffset, UEntAngle, Utrace.HitPos, Angle(math.Clamp(GetConVarNumber("advdupe2_offset_pitch") or 0,-180,180), math.Clamp(GetConVarNumber("advdupe2_offset_yaw") or 0,-180,180), math.Clamp(GetConVarNumber("advdupe2_offset_roll") or 0,-180,180))) 
-			UGhostEnt:SetPos(UPos)
-			UGhostEnt:SetAngles(UAngle)
-		end
 	end
 
 	--Checks binds to modify dupes position and angles
 	function TOOL:Think()
 
-		if(AdvDupe2.HeadGhost)then UpdateGhost() end
-		
+		if AdvDupe2.HeadGhost then
+			AdvDupe2.UpdateGhost()
+		end
+
 		if(LocalPlayer():KeyDown(IN_USE))then
 			if(not AdvDupe2.Rotation)then
 				hook.Add("PlayerBindPress", "AdvDupe2_BindPress", MouseWheelScrolled)
@@ -995,15 +904,15 @@ if(CLIENT)then
 				hook.Remove("PlayerBindPress", "AdvDupe2_BindPress")
 				hook.Remove("CreateMove", "AdvDupe2_MouseControl")
 			end
-			
+
 			XTotal = 0
 			YTotal = 0
 			LastXDegree = 0
-			
+
 			return
 		end
 	end
-	
+
 	--Hinder the player from looking to modify offsets with the mouse
 	function TOOL:FreezeMovement()
 		return AdvDupe2.Rotation
@@ -1017,7 +926,7 @@ if(CLIENT)then
 	language.Add( "Cleanup.AdvDupe2",	"Adv. Duplications" )
 	language.Add( "Cleaned.AdvDupe2",	"Cleaned up all Adv. Duplications" )
 	language.Add( "SBoxLimit.AdvDupe2",	"You've reached the Adv. Duplicator limit!" )
-	
+
 	CreateClientConVar("advdupe2_offset_world", 0, false, true)
 	CreateClientConVar("advdupe2_offset_z", 0, false, true)
 	CreateClientConVar("advdupe2_offset_pitch", 0, false, true)
@@ -1036,7 +945,7 @@ if(CLIENT)then
 	CreateClientConVar("advdupe2_auto_save_contraption", 0, false, true)
 	CreateClientConVar("advdupe2_auto_save_overwrite", 1, false, true)
 	CreateClientConVar("advdupe2_auto_save_time", 10, false, true)
-	
+
 	--Contraption Spawner
 	CreateClientConVar("advdupe2_contr_spawner_key", -1, false, true)
 	CreateClientConVar("advdupe2_contr_spawner_undo_key", -1, false, true)
@@ -1046,92 +955,92 @@ if(CLIENT)then
 	CreateClientConVar("advdupe2_contr_spawner_disdrag", 0, false, true)
 	CreateClientConVar("advdupe2_contr_spawner_addvel", 1, false, true)
 	CreateClientConVar("advdupe2_contr_spawner_hideprops", 0, false, true)
-	
+
 	--Experimental
 	CreateClientConVar("advdupe2_paste_disparents", 0, false, true)
 	CreateClientConVar("advdupe2_paste_protectoveride", 1, false, true)
 	CreateClientConVar("advdupe2_debug_openfile", 1, false, true)
-	
+
 	local function BuildCPanel(CPanel)
 		CPanel:ClearControls()
-		
+
 		local FileBrowser = vgui.Create("advdupe2_browser")
 		CPanel:AddItem(FileBrowser)
 		FileBrowser:SetSize(CPanel:GetWide(),405)
 		AdvDupe2.FileBrowser = FileBrowser
-		
+
 		local Check = vgui.Create("DCheckBoxLabel")
-		
+
 		Check:SetText( "Paste at original position" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_original_origin" ) 
+		Check:SetConVar( "advdupe2_original_origin" )
 		Check:SetValue( 0 )
 		Check:SetToolTip("Paste at the position originally copied")
 		CPanel:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Paste with constraints" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_paste_constraints" ) 
+		Check:SetConVar( "advdupe2_paste_constraints" )
 		Check:SetValue( 1 )
 		Check:SetToolTip("Paste with or without constraints")
 		CPanel:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Paste with parenting" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_paste_parents" ) 
+		Check:SetConVar( "advdupe2_paste_parents" )
 		Check:SetValue( 1 )
 		Check:SetToolTip("Paste with or without parenting")
 		CPanel:AddItem(Check)
-		
+
 		local Check_1 = vgui.Create("DCheckBoxLabel")
 		local Check_2 = vgui.Create("DCheckBoxLabel")
-		
+
 		Check_1:SetText( "Unfreeze all after paste" )
 		Check_1:SetDark(true)
-		Check_1:SetConVar( "advdupe2_paste_unfreeze" ) 
+		Check_1:SetConVar( "advdupe2_paste_unfreeze" )
 		Check_1:SetValue( 0 )
-		Check_1.OnChange = function() 
+		Check_1.OnChange = function()
 			if(Check_1:GetChecked() and Check_2:GetChecked())then
 				Check_2:SetValue(0)
 			end
 		end
 		Check_1:SetToolTip("Unfreeze all props after pasting")
 		CPanel:AddItem(Check_1)
-		
+
 		Check_2:SetText( "Preserve frozen state after paste" )
 		Check_2:SetDark(true)
-		Check_2:SetConVar( "advdupe2_preserve_freeze" ) 
+		Check_2:SetConVar( "advdupe2_preserve_freeze" )
 		Check_2:SetValue( 0 )
-		Check_2.OnChange = function() 
+		Check_2.OnChange = function()
 			if(Check_2:GetChecked() and Check_1:GetChecked())then
 				Check_1:SetValue(0)
 			end
 		end
 		Check_2:SetToolTip("Makes props have the same frozen state as when they were copied")
 		CPanel:AddItem(Check_2)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Area copy constrained props outside of box" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_copy_outside" ) 
+		Check:SetConVar( "advdupe2_copy_outside" )
 		Check:SetValue( 0 )
 		Check:SetToolTip("Copy entities outside of the area copy that are constrained to entities insde")
 		CPanel:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "World/Area copy only your own props" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_copy_only_mine" ) 
+		Check:SetConVar( "advdupe2_copy_only_mine" )
 		Check:SetValue( 1 )
 		Check:SetToolTip("Copy entities outside of the area copy that are constrained to entities insde")
 		CPanel:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Sort constraints by their connections" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_sort_constraints" ) 
+		Check:SetConVar( "advdupe2_sort_constraints" )
 		Check:SetValue( GetConVarNumber("advdupe2_sort_constraints") )
 		Check:SetToolTip( "Orders constraints so that they build a rigid constraint system." )
 		CPanel:AddItem(Check)
@@ -1152,7 +1061,7 @@ if(CLIENT)then
 		local func3 = NumSlider.Wang.Panel.OnLoseFocus
 		NumSlider.Wang.Panel.OnLoseFocus = function(txtBox) func3(txtBox) AdvDupe2.StartGhosting() end
 		CPanel:AddItem(NumSlider)
-		
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Area Copy Size:" )
 		NumSlider.Label:SetDark(true)
@@ -1164,13 +1073,13 @@ if(CLIENT)then
 		NumSlider:SetConVar( "advdupe2_area_copy_size" )
 		NumSlider:SetToolTip("Change the size of the area copy")
 		CPanel:AddItem(NumSlider)
-		
+
 		local Category1 = vgui.Create("DCollapsibleCategory")
 		CPanel:AddItem(Category1)
 		Category1:SetLabel("Offsets")
 		Category1:SetExpanded(0)
-		
-		
+
+
 		local parent = FileBrowser:GetParent():GetParent():GetParent():GetParent()
 		--[[Offsets]]--
 		local CategoryContent1 = vgui.Create( "DPanelList" )
@@ -1179,37 +1088,37 @@ if(CLIENT)then
 		CategoryContent1:SetSpacing( 1 )
 		CategoryContent1:SetPadding( 2 )
 		CategoryContent1.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end		--Fix the damned mouse not scrolling when it's over the catagories
-		
+
 		Category1:SetContents( CategoryContent1 )
 
-				
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Height Offset" )
 		NumSlider.Label:SetDark(true)
 		NumSlider:SetMin( 0 )
-		NumSlider:SetMax( 2500 ) 
-		NumSlider:SetDecimals( 0 ) 
+		NumSlider:SetMax( 2500 )
+		NumSlider:SetDecimals( 0 )
 		NumSlider:SetConVar("advdupe2_offset_z")
 		NumSlider:SetToolTip("Change the Z offset of the dupe")
 		CategoryContent1:AddItem(NumSlider)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Use World Angles" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_offset_world" ) 
+		Check:SetConVar( "advdupe2_offset_world" )
 		Check:SetValue( 0 )
 		Check:SetToolTip("Use world angles for the offset instead of the main entity")
 		CategoryContent1:AddItem(Check)
-		
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Pitch Offset" )
 		NumSlider.Label:SetDark(true)
-		NumSlider:SetMin( -180 ) 
-		NumSlider:SetMax( 180 ) 
-		NumSlider:SetDecimals( 0 ) 
+		NumSlider:SetMin( -180 )
+		NumSlider:SetMax( 180 )
+		NumSlider:SetDecimals( 0 )
 		NumSlider:SetConVar("advdupe2_offset_pitch")
 		CategoryContent1:AddItem(NumSlider)
-				
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Yaw Offset" )
 		NumSlider.Label:SetDark(true)
@@ -1218,7 +1127,7 @@ if(CLIENT)then
 		NumSlider:SetDecimals( 0 )
 		NumSlider:SetConVar("advdupe2_offset_yaw")
 		CategoryContent1:AddItem(NumSlider)
-				
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Roll Offset" )
 		NumSlider.Label:SetDark(true)
@@ -1227,7 +1136,7 @@ if(CLIENT)then
 		NumSlider:SetDecimals( 0 )
 		NumSlider:SetConVar("advdupe2_offset_roll")
 		CategoryContent1:AddItem(NumSlider)
-		
+
 		local Btn = vgui.Create("DButton")
 		Btn:SetText("Reset")
 		Btn.DoClick = function()
@@ -1237,14 +1146,14 @@ if(CLIENT)then
 			RunConsoleCommand("advdupe2_offset_roll", 0)
 		end
 		CategoryContent1:AddItem(Btn)
-		
-		
+
+
 		--[[Dupe Information]]--
 		local Category2 = vgui.Create("DCollapsibleCategory")
 		CPanel:AddItem(Category2)
 		Category2:SetLabel("Dupe Information")
 		Category2:SetExpanded(0)
-				
+
 		local CategoryContent2 = vgui.Create( "DPanelList" )
 		CategoryContent2:SetAutoSize( true )
 		CategoryContent2:SetDrawBackground( false )
@@ -1252,63 +1161,63 @@ if(CLIENT)then
 		CategoryContent2:SetPadding( 2 )
 		Category2:SetContents( CategoryContent2 )
 		CategoryContent2.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end
-		
+
 		AdvDupe2.Info = {}
-		
+
 		local lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.File or "File: ")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.File = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Creator or "Creator:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Creator = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Date or "Date:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Date = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Time or "Time:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Time = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Size or "Size:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Size = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Desc or "Desc:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Desc = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Entities or "Entities:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Entities = lbl
-		
+
 		lbl = vgui.Create( "DLabel" )
 		lbl:SetText(AdvDupe2.InfoText.Constraints or "Constraints:")
 		lbl:SetDark(true)
 		CategoryContent2:AddItem(lbl)
 		AdvDupe2.Info.Constraints = lbl
-	
+
 	--[[Contraption Spawner]]--
 		local Category3 = vgui.Create("DCollapsibleCategory")
 		CPanel:AddItem(Category3)
 		Category3:SetLabel("Contraption Spawner")
 		Category3:SetExpanded(0)
-		
+
 		local CategoryContent3 = vgui.Create( "DPanelList" )
 		CategoryContent3:SetAutoSize( true )
 		CategoryContent3:SetDrawBackground( false )
@@ -1316,14 +1225,14 @@ if(CLIENT)then
 		CategoryContent3:SetPadding( 2 )
 		Category3:SetContents( CategoryContent3 )
 		CategoryContent3.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end
-				
+
 		local ctrl = vgui.Create( "CtrlNumPad" )
 		ctrl:SetConVar1( "advdupe2_contr_spawner_key" )
 		ctrl:SetConVar2( "advdupe2_contr_spawner_undo_key" )
 		ctrl:SetLabel1( "Spawn Key")
 		ctrl:SetLabel2( "Undo Key" )
 		CategoryContent3:AddItem(ctrl)
-			
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Spawn Delay" )
 		NumSlider.Label:SetDark(true)
@@ -1340,11 +1249,11 @@ if(CLIENT)then
 		NumSlider:SetDecimals( 1 )
 		NumSlider:SetConVar("advdupe2_contr_spawner_delay")
 		CategoryContent3:AddItem(NumSlider)
-				
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Undo Delay" )
 		NumSlider.Label:SetDark(true)
-		if(game.SinglePlayer())then 
+		if(game.SinglePlayer())then
 			NumSlider:SetMin( 0 )
 			NumSlider:SetMax( 60 )
 		else
@@ -1361,41 +1270,41 @@ if(CLIENT)then
 		NumSlider:SetDecimals( 1 )
 		NumSlider:SetConVar("advdupe2_contr_spawner_undo_delay")
 		CategoryContent3:AddItem(NumSlider)
-				
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Disable gravity for all spawned props" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_contr_spawner_disgrav" ) 
+		Check:SetConVar( "advdupe2_contr_spawner_disgrav" )
 		Check:SetValue( 0 )
 		CategoryContent3:AddItem(Check)
-				
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Disable drag for all spawned props" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_contr_spawner_disdrag" ) 
+		Check:SetConVar( "advdupe2_contr_spawner_disdrag" )
 		Check:SetValue( 0 )
 		CategoryContent3:AddItem(Check)
-				
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Add spawner's velocity to contraption" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_contr_spawner_addvel" ) 
+		Check:SetConVar( "advdupe2_contr_spawner_addvel" )
 		Check:SetValue( 1 )
 		CategoryContent3:AddItem(Check)
-				
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Disable drawing spawner props" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_contr_spawner_hideprops" ) 
+		Check:SetConVar( "advdupe2_contr_spawner_hideprops" )
 		Check:SetValue( 0 )
 		CategoryContent3:AddItem(Check)
-		
+
 		--[[Area Auto Save]]--
 		local Category4 = vgui.Create("DCollapsibleCategory")
 		CPanel:AddItem(Category4)
 		Category4:SetLabel("Area Auto Save")
 		Category4:SetExpanded(0)
-		
+
 		local CategoryContent4 = vgui.Create( "DPanelList" )
 		CategoryContent4:SetAutoSize( true )
 		CategoryContent4:SetDrawBackground( false )
@@ -1403,15 +1312,15 @@ if(CLIENT)then
 		CategoryContent4:SetPadding( 2 )
 		Category4:SetContents( CategoryContent4 )
 		CategoryContent4.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Only copy contraption" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_auto_save_contraption" ) 
+		Check:SetConVar( "advdupe2_auto_save_contraption" )
 		Check:SetValue( 0 )
 		Check:SetToolTip("Only copy a contraption instead of an area")
 		CategoryContent4:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Overwrite File" )
 		Check:SetDark(true)
@@ -1419,7 +1328,7 @@ if(CLIENT)then
 		Check:SetValue( 1 )
 		Check:SetToolTip("Overwrite the file instead of creating a new one everytime")
 		CategoryContent4:AddItem(Check)
-		
+
 		NumSlider = vgui.Create( "DNumSlider" )
 		NumSlider:SetText( "Minutes to Save:" )
 		NumSlider.Label:SetDark(true)
@@ -1429,19 +1338,19 @@ if(CLIENT)then
 		NumSlider:SetConVar( "advdupe2_auto_save_time" )
 		NumSlider:SetToolTip("Interval time to save in minutes")
 		CategoryContent4:AddItem(NumSlider)
-		
+
 		local pnl = vgui.Create("Panel")
 		pnl:SetWide(CPanel:GetWide()-40)
 		pnl:SetTall(75)
 		pnl:SetPos(0, 50)
 		CategoryContent4:AddItem(pnl)
-		
+
 		local label = vgui.Create("DLabel", pnl)
 		label:SetText("Directory: ")
 		label:SizeToContents()
 		label:SetDark(true)
 		label:SetPos(5,7)
-		
+
 		AdvDupe2.AutoSavePath = ""
 		local txtbox = vgui.Create("DTextEntry", pnl)
 		txtbox:SetWide(pnl:GetWide()-100)
@@ -1450,7 +1359,7 @@ if(CLIENT)then
 		txtbox.OnTextChanged = function(self)
 			self:SetValue(AdvDupe2.AutoSavePath)
 		end
-		
+
 		local btn = vgui.Create("DImageButton", pnl)
 		local x, y = txtbox:GetPos()
 		btn:SetPos(x + txtbox:GetWide() + 5, 7)
@@ -1460,7 +1369,7 @@ if(CLIENT)then
 		btn.DoClick = function()
 			local ScrollBar = parent.VBar
 			ScrollBar:AnimateTo(0, 1, 0, 0.2)
-			
+
 			FileBrowser.Submit:SetMaterial("icon16/disk.png")
 			FileBrowser.Submit:SetTooltip("Directory for Area Auto Save")
 			if(FileBrowser.FileName:GetValue()=="Folder_Name...")then
@@ -1469,7 +1378,7 @@ if(CLIENT)then
 			FileBrowser.Desc:SetVisible(true)
 			FileBrowser.Info:SetVisible(false)
 			FileBrowser.FileName:SetVisible(true)
-			FileBrowser.FileName:SelectAllOnFocus(true) 
+			FileBrowser.FileName:SelectAllOnFocus(true)
 			FileBrowser.FileName:OnMousePressed()
 			FileBrowser.FileName:RequestFocus()
 			FileBrowser.Expanding=true
@@ -1481,25 +1390,25 @@ if(CLIENT)then
 					FileBrowser.FileName:SelectAllOnFocus(true)
 					FileBrowser.FileName:OnGetFocus()
 					FileBrowser.FileName:RequestFocus()
-					return 
-				end 
+					return
+				end
 				local desc = FileBrowser.Desc:GetValue()
 				if(desc=="Description...")then desc="" end
-				
+
 				if(not IsValid(FileBrowser.Browser.pnlCanvas.m_pSelectedItem) or FileBrowser.Browser.pnlCanvas.m_pSelectedItem.Derma.ClassName~="advdupe2_browser_folder")then
 					AdvDupe2.Notify("Folder to save Area Auto Save not selected.", NOTIFY_ERROR)
 					return
 				end
-				
+
 				FileBrowser.AutoSaveNode = FileBrowser.Browser.pnlCanvas.m_pSelectedItem
 				txtbox:SetValue(FileBrowser:GetFullPath(FileBrowser.Browser.pnlCanvas.m_pSelectedItem)..name)
 				AdvDupe2.AutoSavePath = txtbox:GetValue()
 				txtbox:SetToolTip(txtbox:GetValue())
 				AdvDupe2.AutoSaveDesc = desc
-				
+
 				FileBrowser:Slide(false)
 				ScrollBar:AnimateTo(ScrollBar.CanvasSize, 1, 0, 0.2)
-				
+
 				RunConsoleCommand("AdvDupe2_SetStage")
 				hook.Add("HUDPaint", "AdvDupe2_DrawSelectionBox", AdvDupe2.DrawSelectionBox)
 			end
@@ -1511,7 +1420,7 @@ if(CLIENT)then
 			end
 			FileBrowser.Desc.OnEnter = FileBrowser.Submit.DoClick
 		end
-				
+
 		btn = vgui.Create("DButton", pnl)
 		btn:SetSize(50, 35)
 		btn:SetPos(pnl:GetWide()/4-10, 30)
@@ -1524,7 +1433,7 @@ if(CLIENT)then
 				hook.Add("HUDPaint", "AdvDupe2_DrawSelectionBox", AdvDupe2.DrawSelectionBox)
 			end
 		end
-				
+
 		btn = vgui.Create("DButton", pnl)
 		btn:SetSize(50, 35)
 		btn:SetPos((pnl:GetWide()/4)*3-40, 30)
@@ -1537,13 +1446,13 @@ if(CLIENT)then
 		end
 		AdvDupe2.OffButton = btn
 
-		
+
 		--[[Experimental Section]]--
 		local Category5 = vgui.Create("DCollapsibleCategory")
 		CPanel:AddItem(Category5)
 		Category5:SetLabel("Experimental Section")
 		Category5:SetExpanded(0)
-		
+
 		local CategoryContent5 = vgui.Create( "DPanelList" )
 		CategoryContent5:SetAutoSize( true )
 		CategoryContent5:SetDrawBackground( false )
@@ -1551,37 +1460,37 @@ if(CLIENT)then
 		CategoryContent5:SetPadding( 2 )
 		Category5:SetContents( CategoryContent5 )
 		CategoryContent5.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Disable parented props physics interaction" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_paste_disparents" ) 
+		Check:SetConVar( "advdupe2_paste_disparents" )
 		Check:SetValue( 0 )
 		CategoryContent5:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Disable Dupe Spawn Protection" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_paste_protectoveride" ) 
+		Check:SetConVar( "advdupe2_paste_protectoveride" )
 		Check:SetValue( 1 )
 		Check:SetToolTip("Check this if you things don't look right after pasting.")
 		CategoryContent5:AddItem(Check)
-		
+
 		Check = vgui.Create("DCheckBoxLabel")
 		Check:SetText( "Open file after Saving" )
 		Check:SetDark(true)
-		Check:SetConVar( "advdupe2_debug_openfile" ) 
+		Check:SetConVar( "advdupe2_debug_openfile" )
 		Check:SetValue( 1 )
 		Check:SetToolTip("Check this if you want your files to be opened after saving them.")
 		CategoryContent5:AddItem(Check)
-		
+
 		--[[Save Map]]--
 		if(LocalPlayer():IsAdmin())then
 			local Category6 = vgui.Create("DCollapsibleCategory")
 			CPanel:AddItem(Category6)
 			Category6:SetLabel("Save Map")
 			Category6:SetExpanded(0)
-			
+
 			local CategoryContent6 = vgui.Create( "DPanelList" )
 			CategoryContent6:SetAutoSize( true )
 			CategoryContent6:SetDrawBackground( false )
@@ -1589,28 +1498,28 @@ if(CLIENT)then
 			CategoryContent6:SetPadding( 2 )
 			Category6:SetContents( CategoryContent6 )
 			CategoryContent6.OnMouseWheeled = function(self, dlta) parent:OnMouseWheeled(dlta) end
-			
+
 			pnl = vgui.Create("Panel")
 			pnl:SetWide(CPanel:GetWide()-40)
 			pnl:SetTall(75)
 			pnl:SetPos(0, 50)
 			CategoryContent6:AddItem(pnl)
-			
+
 			label = vgui.Create("DLabel", pnl)
 			label:SetText("File Name: ")
 			label:SizeToContents()
 			label:SetDark(true)
 			label:SetPos(5,7)
-			
+
 			AdvDupe2.AutoSavePath = ""
-			
+
 			local txtbox2 = vgui.Create("DTextEntry", pnl)
 			txtbox2:SetWide(pnl:GetWide()-100)
 			txtbox2:SetPos(60, 5)
 			txtbox2.OnEnter = function()
-				btn2:DoClick()	
+				btn2:DoClick()
 			end
-								
+
 			local btn2 = vgui.Create("DImageButton", pnl)
 			x, y = txtbox2:GetPos()
 			btn2:SetPos(x + txtbox2:GetWide() + 5, 7)
@@ -1623,7 +1532,7 @@ if(CLIENT)then
 			end
 		end
 	end
-	
+
 	function TOOL.BuildCPanel(panel)
 		panel:ClearControls()
 		panel:AddControl("Header", {
@@ -1641,296 +1550,6 @@ if(CLIENT)then
 		tryToBuild()
 	end
 
-	function AdvDupe2.LoadGhosts(dupe, info, moreinfo, name, preview)
-
-		AdvDupe2.RemoveGhosts()
-		AdvDupe2.Ghosting = true
-		
-		AdvDupe2.GhostToSpawn = {}
-		local count = 0
-		
-		local time
-		local desc
-		local date
-		local creator
-		
-		if(info.ad1)then
-			time = moreinfo["Time"] or ""
-			desc = info["Description"] or ""
-			date = info["Date"] or ""
-			creator = info["Creator"] or ""
-
-			AdvDupe2.HeadEnt = tonumber(moreinfo.Head)
-			local spx,spy,spz = moreinfo.StartPos:match("^(.-),(.-),(.+)$")
-			AdvDupe2.HeadPos = Vector(tonumber(spx) or 0, tonumber(spy) or 0, tonumber(spz) or 0)
-			local z = (tonumber(moreinfo.HoldPos:match("^.-,.-,(.+)$")) or 0)*-1
-			AdvDupe2.HeadZPos = z
-			AdvDupe2.HeadPos.Z = AdvDupe2.HeadPos.Z + z
-
-			local Pos
-			local Ang
-			for k,v in pairs(dupe["Entities"])do
-				
-				if(v.SavedParentIdx)then 
-					if(not v.BuildDupeInfo)then v.BuildDupeInfo = {} end
-					v.BuildDupeInfo.DupeParentID = v.SavedParentIdx
-					Pos = v.LocalPos*1
-					Ang = v.LocalAngle*1
-				else
-					Pos = nil
-					Ang = nil
-				end
-				for i,p in pairs(v.PhysicsObjects)do
-					p.Pos = Pos or (p.LocalPos*1)
-					p.Pos.Z = p.Pos.Z - z
-					p.Angle = Ang or (p.LocalAngle*1)
-					p.LocalPos = nil
-					p.LocalAngle = nil
-				end
-				v.LocalPos = nil
-				v.LocalAngle = nil
-				AdvDupe2.GhostToSpawn[count] = {Model=v.Model, PhysicsObjects=v.PhysicsObjects}
-				if(AdvDupe2.HeadEnt == k)then
-					AdvDupe2.HeadEnt = count
-				end
-				count = count + 1
-			end
-			
-			AdvDupe2.HeadOffset = AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt].PhysicsObjects[0].Pos
-			AdvDupe2.HeadAngle = AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt].PhysicsObjects[0].Angle
-		
-		else
-			time = info["time"]
-			desc = dupe["Description"]
-			date = info["date"]
-			creator = info["name"]
-
-			AdvDupe2.HeadEnt = dupe["HeadEnt"].Index
-			AdvDupe2.HeadZPos = dupe["HeadEnt"].Z
-			AdvDupe2.HeadPos = dupe["HeadEnt"].Pos
-			AdvDupe2.HeadOffset = dupe["Entities"][AdvDupe2.HeadEnt].PhysicsObjects[0].Pos
-			AdvDupe2.HeadAngle = dupe["Entities"][AdvDupe2.HeadEnt].PhysicsObjects[0].Angle
-			
-			for k,v in pairs(dupe["Entities"])do
-				AdvDupe2.GhostToSpawn[count] = {Model=v.Model, PhysicsObjects=v.PhysicsObjects}
-				if(AdvDupe2.HeadEnt == k)then
-					AdvDupe2.HeadEnt = count
-				end
-				count = count + 1
-			end
-		end
-		
-		if(not preview)then
-			AdvDupe2.Info.File:SetText("File: "..name)
-			AdvDupe2.Info.Creator:SetText("Creator: "..creator)
-			AdvDupe2.Info.Date:SetText("Date: "..date)
-			AdvDupe2.Info.Time:SetText("Time: "..time)
-			AdvDupe2.Info.Size:SetText("Size: "..string.NiceSize(tonumber(info.size) or 0))
-			AdvDupe2.Info.Desc:SetText("Desc: "..(desc or ""))
-			AdvDupe2.Info.Entities:SetText("Entities: "..table.Count(dupe["Entities"]))
-			AdvDupe2.Info.Constraints:SetText("Constraints: "..table.Count(dupe["Constraints"]))
-		end
-		
-		AdvDupe2.StartGhosting()
-		AdvDupe2.Preview = preview
-
-	end
-	
-	function AdvDupe2.RemoveGhosts()
-		
-		if(AdvDupe2.Ghosting)then
-			hook.Remove("Tick", "AdvDupe2_SpawnGhosts")
-			AdvDupe2.Ghosting = false 
-			if(not AdvDupe2.BusyBar)then
-				AdvDupe2.RemoveProgressBar()
-			end
-		end
-
-		if(AdvDupe2.GhostEntities)then
-			for k,v in pairs(AdvDupe2.GhostEntities)do
-				if(IsValid(v))then
-					v:Remove()
-				end
-			end
-		end
-
-		if(IsValid(AdvDupe2.HeadGhost))then
-			AdvDupe2.HeadGhost:Remove()
-		end
-		AdvDupe2.HeadGhost = nil
-		AdvDupe2.CurrentGhost = 1
-		AdvDupe2.GhostEntities = nil
-		AdvDupe2.Preview = false
-	end
-	
-	--Creates a ghost from the given entity's table
-	local function MakeGhostsFromTable(EntTable, gParent)
-
-		if(not EntTable)then return end
-		if(not EntTable.Model or EntTable.Model=="" or EntTable.Model[#EntTable.Model-3]~=".")then EntTable.Model="models/error.mdl" end
-
-		local GhostEntity = ClientsideModel(EntTable.Model, RENDERGROUP_TRANSLUCENT)
-		
-		-- If there are too many entities we might not spawn..
-		if not IsValid(GhostEntity) then 
-			AdvDupe2.RemoveGhosts()
-			AdvDupe2.Notify("Too many entities to spawn ghosts", NOTIFY_ERROR)
-			return 
-		end
-		
-		local Phys = EntTable.PhysicsObjects[0]
-		
-		GhostEntity:SetRenderMode( RENDERMODE_TRANSALPHA )	--Was broken, making ghosts invisible
-		GhostEntity:SetColor( Color(255, 255, 255, 150) )
-
-		-- If we're a ragdoll send our bone positions
-		/*if (EntTable.R) then
-			for k, v in pairs( EntTable.PhysicsObjects ) do
-				if(k==0)then
-					GhostEntity:SetNetworkedBonePosition( k, Vector(0,0,0), v.Angle )
-				else
-					GhostEntity:SetNetworkedBonePosition( k, v.Pos, v.Angle )
-				end
-			end	
-			Phys.Angle = Angle(0,0,0)
-		end*/
-		
-		if ( gParent ) then
-			local Parent = AdvDupe2.HeadGhost
-			local temp = Parent:GetAngles()
-			GhostEntity:SetPos(Parent:GetPos() + Phys.Pos - AdvDupe2.HeadOffset)
-			GhostEntity:SetAngles(Phys.Angle)
-			Parent:SetAngles(AdvDupe2.HeadAngle)
-			GhostEntity:SetParent(Parent)
-			Parent:SetAngles(temp)
-		else
-			GhostEntity:SetAngles(Phys.Angle)
-		end
-		
-		return GhostEntity
-	end
-	
-	local gTemp = 0
-	local gPerc = 0
-	local function SpawnGhosts()
-		AdvDupe2.CurrentGhost = AdvDupe2.CurrentGhost + math.floor(gPerc)
-		gTemp = gTemp + gPerc - math.floor(gPerc)
-		if(gTemp>1)then
-			AdvDupe2.CurrentGhost = AdvDupe2.CurrentGhost + 1
-			gTemp = gTemp - math.floor(gTemp)
-		end
-		if(AdvDupe2.CurrentGhost==AdvDupe2.HeadEnt)then AdvDupe2.CurrentGhost = AdvDupe2.CurrentGhost + 1 end
-		
-		local g = AdvDupe2.GhostToSpawn[AdvDupe2.CurrentGhost]
-		if(not g)then
-			AdvDupe2.Ghosting = false
-			hook.Remove("Tick", "AdvDupe2_SpawnGhosts")
-			if(not AdvDupe2.BusyBar)then
-				AdvDupe2.RemoveProgressBar()
-			end
-			return
-		end
-		AdvDupe2.GhostEntities[AdvDupe2.CurrentGhost] = MakeGhostsFromTable(g, true)
-		if(not AdvDupe2.BusyBar)then
-			AdvDupe2.ProgressBar.Percent = AdvDupe2.CurrentGhost/AdvDupe2.TotalGhosts*100
-		end
-	end
-	
-	net.Receive("AdvDupe2_SendGhosts", 	function(len, ply, len2)
-		AdvDupe2.RemoveGhosts()
-		AdvDupe2.Ghosting = true
-		AdvDupe2.GhostToSpawn = {}
-		AdvDupe2.HeadEnt = net.ReadInt(16)
-		AdvDupe2.HeadZPos = net.ReadFloat()
-		AdvDupe2.HeadPos = net.ReadVector()
-		local cache = {}
-		for i=1, net.ReadInt(16) do
-			cache[i] = net.ReadString()
-		end
-		
-		for i=1, net.ReadInt(16) do
-			AdvDupe2.GhostToSpawn[i] = {R = net.ReadBit()==1, Model = cache[net.ReadInt(16)], PhysicsObjects = {}}
-			for k=0, net.ReadInt(8) do
-				AdvDupe2.GhostToSpawn[i].PhysicsObjects[k] = {Angle = net.ReadAngle(), Pos = net.ReadVector()}
-			end
-		end
-		AdvDupe2.GhostEntities = {}
-		AdvDupe2.HeadGhost = MakeGhostsFromTable(AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt])
-		AdvDupe2.HeadOffset = AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt].PhysicsObjects[0].Pos
-		AdvDupe2.HeadAngle = AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt].PhysicsObjects[0].Angle
-		AdvDupe2.GhostEntities[AdvDupe2.HeadEnt] = AdvDupe2.HeadGhost	
-		AdvDupe2.CurrentGhost = 0
-		AdvDupe2.TotalGhosts = table.Count(AdvDupe2.GhostToSpawn)
-		
-		if(AdvDupe2.TotalGhosts>1)then
-			gTemp = 0
-			gPerc = AdvDupe2.TotalGhosts*(GetConVarNumber("advdupe2_limit_ghost")*0.01)
-			if(gPerc>0)then
-				gPerc = AdvDupe2.TotalGhosts / gPerc
-				if(not AdvDupe2.BusyBar)then
-					AdvDupe2.InitProgressBar("Ghosting: ")
-					AdvDupe2.BusyBar = false
-				end
-				hook.Add("Tick", "AdvDupe2_SpawnGhosts", SpawnGhosts)
-			else
-				AdvDupe2.Ghosting = false
-			end
-		else
-			AdvDupe2.Ghosting = false
-		end
-	end)
-										
-	net.Receive("AdvDupe2_AddGhost", 	function(len, ply, len2)
-		local gNew = table.insert(AdvDupe2.GhostToSpawn, {R = net.ReadBit()==1, Model = net.ReadString(), PhysicsObjects = {}})
-		for k=0, net.ReadInt(8) do
-			AdvDupe2.GhostToSpawn[gNew].PhysicsObjects[k] = {Angle = net.ReadAngle(), Pos = net.ReadVector()}
-		end
-		
-		if(AdvDupe2.CurrentGhost==gNew)then
-			AdvDupe2.GhostEntities[gNew] = MakeGhostsFromTable(AdvDupe2.GhostToSpawn[gNew], true)
-			AdvDupe2.CurrentGhost = AdvDupe2.CurrentGhost + math.floor(gPerc)
-			gTemp = gTemp + gPerc - math.floor(gPerc)
-			if(gTemp>1)then
-				AdvDupe2.CurrentGhost = AdvDupe2.CurrentGhost + 1
-				gTemp = gTemp - math.floor(gTemp)
-			end
-		end
-	end)
-
-	function AdvDupe2.StartGhosting()
-		AdvDupe2.RemoveGhosts()
-		if(not AdvDupe2.GhostToSpawn)then return end
-		AdvDupe2.Ghosting = true
-		AdvDupe2.GhostEntities = {}
-		AdvDupe2.HeadGhost = MakeGhostsFromTable(AdvDupe2.GhostToSpawn[AdvDupe2.HeadEnt])
-		AdvDupe2.GhostEntities[AdvDupe2.HeadEnt] = AdvDupe2.HeadGhost
-		AdvDupe2.CurrentGhost = 0
-		AdvDupe2.TotalGhosts = #AdvDupe2.GhostToSpawn
-		
-		if(AdvDupe2.TotalGhosts  > 1)then
-			gTemp = 0
-			gPerc = AdvDupe2.TotalGhosts*(GetConVarNumber("advdupe2_limit_ghost")*0.01) - 1
-			if(gPerc>0)then
-				gPerc = AdvDupe2.TotalGhosts / gPerc
-				if(not AdvDupe2.BusyBar)then
-					AdvDupe2.InitProgressBar("Ghosting: ")
-					AdvDupe2.BusyBar = false
-				end
-				hook.Add("Tick", "AdvDupe2_SpawnGhosts", SpawnGhosts)
-			else
-				AdvDupe2.Ghosting = false
-			end
-		else
-			AdvDupe2.Ghosting = false
-		end
-	end
-	usermessage.Hook("AdvDupe2_StartGhosting", function()
-		AdvDupe2.StartGhosting()
-	end)
-												
-	usermessage.Hook("AdvDupe2_RemoveGhosts", AdvDupe2.RemoveGhosts)
-	
-
 	local state = 0
 	local ToColor = {r=25, g=100, b=40, a=255}
 	local CurColor = {r=25, g=100, b=40, a=255}
@@ -1939,7 +1558,7 @@ if(CLIENT)then
 	surface.CreateFont ("AD2TitleFont", {font="Arial", size=24, weight=1000})
 	function TOOL:DrawToolScreen()
 		if(not AdvDupe2)then return true end
-		
+
 		local text = "Ready"
 		if(AdvDupe2.Preview)then
 			text = "Preview"
@@ -1949,25 +1568,25 @@ if(CLIENT)then
 			state=1
 			text = AdvDupe2.ProgressBar.Text
 		end
-		
+
 		cam.Start2D()
 
 			surface.SetDrawColor(32, 32, 32, 255)
 			surface.DrawRect(0, 0, 256, 256)
-			
+
 			if(state==0)then
 				ToColor = {r=25, g=100, b=40, a=255}
 			else
 				ToColor = {r=130, g=25, b=40, a=255}
 			end
-			
+
 			rate = FrameTime()*160
 			CurColor.r = math.Approach( CurColor.r, ToColor.r, rate )
 			CurColor.g = math.Approach( CurColor.g, ToColor.g, rate )
-			
+
 			surface.SetDrawColor(CurColor)
 			surface.DrawRect(13, 13, 230, 230)
-			
+
 			surface.SetTextColor( 255, 255, 255, 255 )
 
 			draw.SimpleText("Advanced Duplicator 2", "AD2TitleFont", 128, 50, Color(255,255,255,255), TEXT_ALIGN_CENTER, TEXT_ALIGN_TOP)
@@ -1980,11 +1599,11 @@ if(CLIENT)then
 				draw.SimpleText("Pitch: "..LocalPlayer():GetInfo("advdupe2_offset_pitch"), "AD2TitleFont", 25, 190, Color(255,255,255,255), TEXT_ALIGN_LEFT, TEXT_ALIGN_BOTTOM)
 				draw.SimpleText("Yaw: "..LocalPlayer():GetInfo("advdupe2_offset_yaw"), "AD2TitleFont", 25, 220, Color(255,255,255,255), TEXT_ALIGN_LEFT, TEXT_ALIGN_BOTTOM)
 			end
-			
+
 		cam.End2D()
 	end
 
-	
+
 	local function FindInBox(min, max, ply)
 
 		local Entities = ents.GetAll()
@@ -2000,28 +1619,28 @@ if(CLIENT)then
 
 		return EntTable
 	end
-	
-	
+
+
 	local GreenSelected = Color(0, 255, 0, 255)
 	function AdvDupe2.DrawSelectionBox()
-			
+
 		local TraceRes = util.TraceLine(util.GetPlayerTrace(LocalPlayer()))
 		local i = tonumber(LocalPlayer():GetInfo("advdupe2_area_copy_size")) or 50
-				
+
 		--Bottom Points
 		local B1 = (Vector(-i,-i,-i)+TraceRes.HitPos)
 		local B2 = (Vector(-i,i,-i)+TraceRes.HitPos)
 		local B3 = (Vector(i,i,-i)+TraceRes.HitPos)
 		local B4 = (Vector(i,-i,-i)+TraceRes.HitPos)
-				
+
 		--Top Points
 		local T1 = (Vector(-i,-i,i)+TraceRes.HitPos):ToScreen()
 		local T2 = (Vector(-i,i,i)+TraceRes.HitPos):ToScreen()
 		local T3 = (Vector(i,i,i)+TraceRes.HitPos):ToScreen()
 		local T4 = (Vector(i,-i,i)+TraceRes.HitPos):ToScreen()
-				
+
 		if(not AdvDupe2.LastUpdate or CurTime()>=AdvDupe2.LastUpdate)then
-			
+
 			if AdvDupe2.ColorEntities then
 				for k,v in pairs(AdvDupe2.EntityColors)do
 					local ent = AdvDupe2.ColorEntities[k]
@@ -2030,7 +1649,7 @@ if(CLIENT)then
 					end
 				end
 			end
-					
+
 			local Entities = FindInBox(B1, (Vector(i,i,i)+TraceRes.HitPos), LocalPlayer())
 			AdvDupe2.ColorEntities = Entities
 			AdvDupe2.EntityColors = {}
@@ -2039,13 +1658,13 @@ if(CLIENT)then
 				v:SetColor(GreenSelected)
 			end
 			AdvDupe2.LastUpdate = CurTime()+0.25
-				
+
 		end
-				
+
 		local tracedata = {}
 		tracedata.mask = MASK_NPCWORLDSTATIC
 		local WorldTrace
-				
+
 		tracedata.start = B1+Vector(0,0,i*2)
 		tracedata.endpos = B1
 		WorldTrace = util.TraceLine( tracedata )
@@ -2062,21 +1681,21 @@ if(CLIENT)then
 		tracedata.endpos = B4
 		WorldTrace = util.TraceLine( tracedata )
 		B4 = WorldTrace.HitPos:ToScreen()
-				
+
 		surface.SetDrawColor( 0, 255, 0, 255 )
-				
+
 		--Draw Sides
 		surface.DrawLine(B1.x, B1.y, T1.x, T1.y)
 		surface.DrawLine(B2.x, B2.y, T2.x, T2.y)
 		surface.DrawLine(B3.x, B3.y, T3.x, T3.y)
 		surface.DrawLine(B4.x, B4.y, T4.x, T4.y)
-				
+
 		--Draw Bottom
 		surface.DrawLine(B1.x, B1.y, B2.x, B2.y)
 		surface.DrawLine(B2.x, B2.y, B3.x, B3.y)
 		surface.DrawLine(B3.x, B3.y, B4.x, B4.y)
 		surface.DrawLine(B4.x, B4.y, B1.x, B1.y)
-			
+
 		--Draw Top
 		surface.DrawLine(T1.x, T1.y, T2.x, T2.y)
 		surface.DrawLine(T2.x, T2.y, T3.x, T3.y)
@@ -2084,16 +1703,16 @@ if(CLIENT)then
 		surface.DrawLine(T4.x, T4.y, T1.x, T1.y)
 
 	end
-			
-	usermessage.Hook("AdvDupe2_DrawSelectBox",function() 
-		hook.Add("HUDPaint", "AdvDupe2_DrawSelectionBox", AdvDupe2.DrawSelectionBox) 
+
+	usermessage.Hook("AdvDupe2_DrawSelectBox",function()
+		hook.Add("HUDPaint", "AdvDupe2_DrawSelectionBox", AdvDupe2.DrawSelectionBox)
 	end)
-	
+
 	function AdvDupe2.RemoveSelectBox()
-		hook.Remove("HUDPaint", "AdvDupe2_DrawSelectionBox") 
+		hook.Remove("HUDPaint", "AdvDupe2_DrawSelectionBox")
 		if AdvDupe2.ColorEntities then
 			for k,v in pairs(AdvDupe2.EntityColors)do
-				if(not IsValid(AdvDupe2.ColorEntities[k]))then 
+				if(not IsValid(AdvDupe2.ColorEntities[k]))then
 					AdvDupe2.ColorEntities[k]=nil
 				else
 					AdvDupe2.ColorEntities[k]:SetColor(v)
@@ -2103,10 +1722,10 @@ if(CLIENT)then
 			AdvDupe2.EntityColors={}
 		end
 	end
-	usermessage.Hook("AdvDupe2_RemoveSelectBox",function() 
+	usermessage.Hook("AdvDupe2_RemoveSelectBox",function()
 		AdvDupe2.RemoveSelectBox()
 	end)
-	
+
 	function AdvDupe2.InitProgressBar(label)
 		AdvDupe2.ProgressBar = {}
 		AdvDupe2.ProgressBar.Text = label
@@ -2120,7 +1739,7 @@ if(CLIENT)then
 	usermessage.Hook("AdvDupe2_UpdateProgressBar",function(um)
 		AdvDupe2.ProgressBar.Percent = um:ReadChar()
 	end)
-	
+
 	function AdvDupe2.RemoveProgressBar()
 		AdvDupe2.ProgressBar = {}
 		AdvDupe2.BusyBar = false
@@ -2133,7 +1752,7 @@ if(CLIENT)then
 	usermessage.Hook("AdvDupe2_RemoveProgressBar",function(um)
 		AdvDupe2.RemoveProgressBar()
 	end)
-	
+
 	usermessage.Hook("AdvDupe2_ResetOffsets",function(um)
 		RunConsoleCommand("advdupe2_original_origin", "0")
 		RunConsoleCommand("advdupe2_paste_constraints","1")
@@ -2144,15 +1763,15 @@ if(CLIENT)then
 		RunConsoleCommand("advdupe2_paste_parents","1")
 		RunConsoleCommand("advdupe2_paste_disparents","0")
 	end)
-	
+
 	usermessage.Hook("AdvDupe2_ReportModel",function(um)
 		print("Advanced Duplicator 2: Invalid Model: "..um:ReadString())
 	end)
-	
+
 	usermessage.Hook("AdvDupe2_ReportClass",function(um)
 		print("Advanced Duplicator 2: Invalid Class: "..um:ReadString())
 	end)
-	
+
 	usermessage.Hook("AdvDupe2_ResetDupeInfo", function(um)
 		if not AdvDupe2.Info then return end
 		AdvDupe2.Info.File:SetText("File:")
@@ -2164,7 +1783,7 @@ if(CLIENT)then
 		AdvDupe2.Info.Entities:SetText("Entities:")
 		AdvDupe2.Info.Constraints:SetText("Constraints:")
 	end)
-	
+
 	usermessage.Hook("AdvDupe2_CanAutoSave", function(um)
 		if(AdvDupe2.AutoSavePath~="")then
 			AdvDupe2.AutoSavePos = um:ReadVector()
@@ -2182,7 +1801,7 @@ if(CLIENT)then
 			AdvDupe2.Notify("Select a directory for the Area Auto Save.", NOTIFY_ERROR)
 		end
 	end)
-	
+
 	net.Receive("AdvDupe2_SetDupeInfo", function(len, ply, len2)
 		if AdvDupe2.Info then
 			AdvDupe2.Info.File:SetText("File: "..net.ReadString())
